+++ src/extension.ts
@@ -39,3 +39,4 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 export function deactivate() {}
+