+++ src/test/components/typeInlineDeclaration/typeInlineDeclaration.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function Component() {
+  const anotherClass: string = 'my-class';
+  const baseClass: string = 'my-class-2';
+
+  return (
+    <div className={baseClass} style={{ color: 'red' }}>
+      <div className={anotherClass}>Test</div>
+    </div>
+  );
+}
+