+++ src/test/components/undestructuredPropsSpreadAttribute/undestructuredPropsSpreadAttribute.tsx
@@ -1,0 +1,10 @@
+import React, { ComponentPropsWithoutRef } from 'react';
+
+function Component({ children, ...props }: ComponentPropsWithoutRef<'div'>) {
+  return (
+    <div>
+      <div {...props}>{children}</div>
+    </div>
+  );
+}
+