+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [0.9.0] - 2024-06-28
+
+### Added
+
+- Configuration option for inline props type declaration.
+- Configuration option for keeping props undestructured.
+
+### Fixed
+
+- Description for "Explicit Return Statement" configuration.
+
 ## [0.8.1] - 2024-05-18
 
 ### Fixed